NAS-XXXXXX: Preserve pool disks for Disk Temperature Reports
