+++ src/app/pages/storage/components/disk-health-card/disk-health-card.component.html
@@ -50,7 +50,7 @@
           {{ diskState.symbolText }}{{ diskState.unit }}
         </ng-container>
       </div>
-      <a [routerLink]="['/reportsdashboard/disk']">{{ 'View Reports' | translate }}</a>
+      <a [routerLink]="['/reportsdashboard/disk']" [queryParams]="{ disks: allDiskNames }">{{ 'View Reports' | translate }}</a>
     </div>
     <div class="details-item">
       <div class="label">{{ 'Failed S.M.A.R.T. Tests' | translate }}:</div>
